Extracted Addons methods to a separate file
